+++ src/main/java/com/dionialves/snapdogdelivery/order/OrderResponseDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -26,6 +27,6 @@ public class OrderResponseDTO {
     @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate date;
 
-    private Double totalValue;
+    private BigDecimal totalValue;
 
 }
